biblioteca/core/custom_permissions.py
from rest_framework import permissions

class IsCollectionOwnerOrReadOnly(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        # Allow read permissions for any request
        if request.method in permissions.SAFE_METHODS:
            return True

        # Write permissions are only allowed to the owner of the collection
        return obj.colecionador == request.user
