biblioteca/requirements.txt
Django==1.11.29
djangorestframework==3.9.4
django-filter==1.1.0
pytz==2024.2
